+++ Tabloid/client/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router } from "react-router-dom";
 import { UserProfileProvider } from "./providers/UserProfileProvider";
-import Header from "./components/Header";
-import ApplicationViews from "./components/ApplicationViews";
+import Header from "./components/layout/Header";
+import ApplicationViews from "./components/layout/ApplicationViews";
 
 function App() {
   return (
