+++ Tabloid/client/src/components/layout/ApplicationViews.js
@@ -4,8 +4,8 @@ import {
   UserProfileContext,
   UserProfileProvider,
 } from "../../providers/UserProfileProvider";
-import Login from "./Login";
-import Register from "./Register";
+import Login from "../auth/Login";
+import Register from "../auth/Register";
 import Hello from "./Hello";
 
 export default function ApplicationViews() {
